Move mersenne twister to random package
